Завершил рефакторинг. Перенес код из майн в презентеры
